# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Transaction Service API
  version: 1.0.0

servers:
  - url: http://localhost/api/v1

paths:
  /transactions/account:
    post:
      summary: Create transaction for accounts
      operationId: createTransactionAccounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountTransactionRequest'
      responses:
        '201':
          description: Account transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountTransactionResponse'

  /transactions/account/by-account/{accountNumber}:
    get:
      summary: Get account transactions by account
      operationId: getAccountTransactionsByAccountNumber
      parameters:
        - name: accountNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccountTransactionResponse'

  /transactions/credit-card/by-card-number/{creditCardNumber}:
    get:
      summary: Get credit card transactions by credit card number
      operationId: getCreditCardTransactionsByCreditCardNumber
      parameters:
        - name: creditCardNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditCardTransactionResponse'

  /transactions/credit/by-credit/{creditId}:
    get:
      summary: Get credit transactions by credit id
      operationId: getCreditTransactionsByCreditId
      parameters:
        - name: creditId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreditPaymentTransactionResponse'

  /transactions/credit-card:
    post:
      summary: Create a transaction for credit cards
      operationId: createCreditCardTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardTransactionRequest'
      responses:
        '201':
          description: Credit card transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardTransactionResponse'

  /transactions/debit-card:
    post:
      summary: Create a transaction for debit cards
      operationId: createDebitCardTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DebitCardTransactionRequest'
      responses:
        '201':
          description: Debit card transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DebitCardTransactionResponse'

  /transactions/credit/payment:
    post:
      summary: Create a payment transaction for credit
      operationId: createCreditPaymentTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPaymentTransactionRequest'
      responses:
        '201':
          description: Credit payment transaction created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditPaymentTransactionResponse'

  /transactions/customer-products/{customerId}:
    get:
      summary: Get all the customer bank products
      operationId: getAllCustomerProductsByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all customer products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProductsResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /transactions/customer-products/avg-balance/{customerId}:
    get:
      summary: Get all the customer bank products average balance
      operationId: getAllCustomerProductsAvgBalanceByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all customer products with its avg monthly balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProductsAverageBalanceResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

  /transactions/customer-products/dni/{dni}:
    get:
      summary: Get all the customer bank products
      operationId: getAllCustomerProductsByDni
      parameters:
        - name: dni
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all customer products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProductsResponse'

  /transactions/customer-products/ruc/{ruc}:
    get:
      summary: Get all the customer bank products
      operationId: getAllCustomerProductsByRuc
      parameters:
        - name: ruc
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of all customer products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerProductsResponse'

  /transactions/products/commissions/range:
    get:
      summary: Get all the products commission between a date range
      operationId: getAllProductsCommissionRange
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of all products commission between a date range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductsCommissionResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'

components:
  schemas:
    CustomerProductsResponse:
      type: object
      properties:
        customer:
          $ref: "#/components/schemas/CustomerResponse"
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/AccountResponse"
        creditCards:
          type: array
          items:
            $ref: "#/components/schemas/CreditCardResponse"
        credits:
          type: array
          items:
            $ref: "#/components/schemas/CreditResponse"

    AccountTransactionRequest:
      type: object
      properties:
        originAccountNumber:
          type: string
          minLength: 1
          maxLength: 40
        destinationAccountNumber:
          type: string
          minLength: 1
          maxLength: 40
        type:
          type: string
          minLength: 1
        amount:
          type: number
        description:
          type: string
      required:
        - originAccountNumber
        - type
        - amount

    CreditCardTransactionRequest:
      type: object
      properties:
        creditCard:
          type: string
        customerId:
          type: string
        amount:
          type: number
          minimum: 1
        description:
          type: string
        type:
          type: string
          x-field-extra-annotation: '@com.project1.ms_transaction_service.validation.EnumNamePattern(regexp = "USAGE|PAYMENT")'
      required:
        - creditCard
        - customerId
        - amount
        - type

    CreditCardTransactionResponse:
      type: object
      properties:
        creditCard:
          type: string
        customerId:
          type: string
        amount:
          type: number
        description:
          type: string
        type:
          type: string
        date:
          type: string
          format: date-time

    DebitCardTransactionRequest:
      type: object
      properties:
        debitCardId:
          type: string
        amount:
          type: number
          minimum: 1
        description:
          type: string
        type:
          type: string
          x-field-extra-annotation: '@com.project1.ms_transaction_service.validation.EnumNamePattern(regexp = "PAYMENT|WITHDRAWAL")'
      required:
        - debitCardId
        - amount
        - type

    DebitCardTransactionResponse:
      type: object
      properties:
        debitCardId:
          type: string
        customerId:
          type: string
        amount:
          type: number
        description:
          type: string
        type:
          type: string
        date:
          type: string
          format: date-time

    DebitCardResponse:
      type: object
      properties:
        cardNumber:
          type: string
        customerId:
          type: string
        associations:
          type: array
          items:
            $ref: '#/components/schemas/DebitCardAssociation'

    DebitCardAssociation:
      type: object
      properties:
        accountId:
          type: string
        position:
          type: integer

    CreditPaymentTransactionRequest:
      type: object
      properties:
        creditId:
          type: string
      required:
        - creditId
        - customerId
        - amount

    CreditPaymentTransactionResponse:
      type: object
      properties:
        creditId:
          type: string
        date:
          type: string
          format: date-time

    CreditPatchRequest:
      type: object
      properties:
        amountPaid:
          type: number
      required:
        - amountPaid

    AccountTransactionResponse:
      type: object
      properties:
        id:
          type: string
        originAccountNumber:
          type: string
        destinationAccountNumber:
          type: string
        type:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date-time
        description:
          type: string

    AccountPatchRequest:
      type: object
      properties:
        status:
          type: string
        balance:
          type: number
        monthlyMovements:
          type: integer

    CreditCardPatchRequest:
      type: object
      properties:
        usedAmount:
          type: number
        creditLimit:
          type: number

    AccountResponse:
      type: object
      properties:
        id:
          type: string
        accountNumber:
          type: string
        accountType:
          type: string
        balance:
          type: number
        customerId:
          type: string
        creationDate:
          type: string
          format: date-time
        status:
          type: string
        monthlyMovements:
          type: integer
        maintenanceFee:
          type: number
          format: double
        endDay:
          type: string
          format: date-time
        availableDayForMovements:
          type: integer
        maxMonthlyMovements:
          type: integer
        customerType:
          type: string
        maxMonthlyMovementsNoFee:
          type: integer
        transactionCommissionFeePercentage:
          type: number

    CreditCardResponse:
      type: object
      properties:
        id:
          type: string
        customerId:
          type: string
        creditLimit:
          type: number
        cardNumber:
          type: string
        usedAmount:
          type: number
        creditCardStatus:
          type: string

    CreditResponse:
      type: object
      properties:
        clientId:
          type: string
        amount:
          type: number
        status:
          type: string
        interestRate:
          type: number
        termInMonths:
          type: integer
        monthlyPayment:
          type: number
        amountPaid:
          type: number
        totalAmount:
          type: number
        identifier:
          type: string

    CustomerResponse:
      type: object
      properties:
        id:
          type: string
        documentNumber:
          type: string
        ruc:
          type: string
        businessName:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        type:
          type: string
        status:
          type: string

    ResponseBase:
      type: object
      properties:
        message:
          type: string

    # queries
    CustomerProductsAverageBalanceResponse:
      type: object
      properties:
        customer:
          $ref: '#/components/schemas/CustomerResponse'
        accounts:
          type: array
          items:
            type: object
            properties:
              accountNumber:
                type: string
              averageBalance:
                type: number
        creditCards:
          type: array
          items:
            type: object
            properties:
              cardNumber:
                type: string
              averageBalance:
                type: number
        credits:
          type: array
          items:
            type: object
            properties:
              creditIdentifier:
                type: string
              averageBalance:
                type: number

    ProductsCommissionResponse:
      type: object
      properties:
        commissions:
          type: object
          properties:
            accounts:
              type: array
              items:
                type: object
                properties:
                  accountNumber:
                    type: string
                  totalCommissionFee:
                    type: number
            totalAccountsCommissionFee:
              type: number